+++ src/node_gdal.cpp
@@ -194,7 +194,7 @@ static NAN_METHOD(Log) {
 }
 
 /*
- * Common code for sync and async opening
+ * Common code for sync and async opening.
  */
 GDAL_ASYNCABLE_GLOBAL(gdal_open);
 GDAL_ASYNCABLE_DEFINE(gdal_open) {
@@ -267,7 +267,11 @@ static NAN_METHOD(getConfigOption) {
 }
 
 /**
- * Convert decimal degrees to degrees, minutes, and seconds string
+ * @typedef {string[]|Record<string, string|number|(string|number)[]>} StringOptions
+ */
+
+/**
+ * Convert decimal degrees to degrees, minutes, and seconds string.
  *
  * @static
  * @method decToDMS
@@ -277,7 +281,6 @@ static NAN_METHOD(getConfigOption) {
  * @return {string} A string nndnn'nn.nn'"L where n is a number and L is either
  * N or E
  */
-
 static NAN_METHOD(decToDMS) {
 
   double angle;