+++ test/api_base.test.ts
@@ -11,7 +11,8 @@ if (process.env.GDAL_DATA !== undefined) {
 }
 
 describe('gdal', () => {
-  afterEach(global.gc)
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  afterEach(global.gc!)
 
   describe('"lastError" property', () => {
     describe('get()', () => {
@@ -73,11 +74,9 @@ describe('gdal', () => {
         path.resolve(__dirname, '..', 'deps', 'libgdal', 'gdal', 'data') :
         path.resolve(__dirname, '..', 'node_modules', 'gdal-async', 'deps', 'libgdal', 'gdal', 'data')
 
-      if (gdal.bundled) {
-        it('should set GDAL_DATA config option to locally bundled path', () => {
-          assert.equal(gdal.config.get('GDAL_DATA'), data_path)
-        })
-      }
+      it('with bundled GDAL, should set GDAL_DATA config option to locally bundled path', () => {
+        assert.equal(gdal.config.get('GDAL_DATA'), data_path)
+      })
       it('should respect GDAL_DATA environment over locally bundled path', (done) => {
         const env = Object.assign({}, process.env)
         env.GDAL_DATA = 'bogus'