+++ test/api_drivers.test.ts
@@ -6,7 +6,8 @@ import * as gdal from 'gdal-async'
 chai.use(chaiAsPromised)
 
 describe('gdal.drivers', () => {
-  afterEach(global.gc)
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  afterEach(global.gc!)
 
   describe('count()', () => {
     it('should return the number of drivers', () => {
@@ -160,7 +161,41 @@ describe('gdal.drivers', () => {
     })
   })
 
-  describe('createCopy', () => {
+  describe('open()', () => {
+    it('should operate normally', () => {
+      const driver = gdal.drivers.get('GTiff')
+      const ds = driver.open(`${__dirname}/data/sample.tif`)
+      assert.instanceOf(ds, gdal.Dataset)
+      assert.isNotNull(ds.srs)
+      assert.isNotNull(ds.geoTransform)
+      assert.strictEqual(ds.driver, driver)
+    })
+
+    it('should support passing driver-specific open options', () => {
+      const driver = gdal.drivers.get('GTiff')
+      const ds1 = driver.open(`${__dirname}/data/sample.tif`, 'r', { GEOREF_SOURCES: 'NONE' })
+      assert.instanceOf(ds1, gdal.Dataset)
+      assert.isNull(ds1.srs)
+      assert.isNull(ds1.geoTransform)
+      assert.strictEqual(ds1.driver, driver)
+
+      const ds2 = driver.open(`${__dirname}/data/sample.tif`, 'r', [ 'GEOREF_SOURCES=NONE' ])
+      assert.instanceOf(ds2, gdal.Dataset)
+      assert.isNull(ds2.srs)
+      assert.isNull(ds2.geoTransform)
+      assert.strictEqual(ds2.driver, driver)
+    })
+
+    it('should throw if the driver-specific options are invalid', () => {
+      assert.throws(() => {
+        const driver = gdal.drivers.get('GTiff')
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        driver.open(`${__dirname}/data/sample.tif`, 'r', 'GEOREF_SOURCES=NONE' as any)
+      }, /Failed parsing options/)
+    })
+  })
+
+  describe('createCopy()', () => {
     it('should operate normally', () => {
       const driver = gdal.drivers.get('MEM')
       const outputFilename = '' // __dirname + '/data/12_791_1476.tif';
@@ -197,6 +232,19 @@ describe('gdal.drivers', () => {
           gdal.open(`${__dirname}/data/12_791_1476.jpg`)
         )
       }, /No such file/)
+    })
+
+    it('should throw on invalid options', () => {
+      const driver = gdal.drivers.get('GTiff')
+      const outputFilename = '' // __dirname + '/data/12_791_1476.tif';
+      assert.throws(() => {
+        driver.createCopy(
+          outputFilename,
+          gdal.open(`${__dirname}/data/12_791_1476.jpg`),
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          'option=invalid' as any
+        )
+      }, /Failed parsing options/)
     })
   })